requirements.txt
cycler==0.11.0
dm-sonnet==2.0.0
ipython==7.29.0
jupyter==1.0.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
Pillow==8.4.0
pyparsing==3.0.4
python-dateutil==2.8.2
scipy==1.7.1
six==1.16.0
tensorflow==2.7.0
tensorflow-probability==0.15.0
torch==1.10.0
torchvision==0.11.1
typing-extensions==3.10.0.2
